+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+# Version 1.0.1
+- Duplicate states are now marked.
+- You can now toggle checkerboard background
+- Copy/Paste/Cut fixes. (Hotkeys only at the moment, context menu pending)
+- Copied states/frames should now be accessible in external editors.
+- Added copying pngs from external editors/filesystem as frame/state.
+- Shrinking dmis produces correct metadata now.
+
 # Version 1.0.0
 - First release.